Launch function & prep for unittest
Wraps main game script in a function.
Function can be deployed with mode='default' for normal gameplay, or mode='diag' for testing.
Intro_players function is wrapped with a decorator to bypass prompts in 'diag' mode.
Load_defaults function is wrapped with a decorator (INCOMPLETE) which accepts mode AND a set of kwargs which will be used to specify the current state of the game in unittest scripts.
The remaining steps to complete this feature are: (1) in the load_defaults decorator, count the occurences of each marker (from a custom board state, kwarg 'board') to derive who the "current" player is. Then execute the win_check function and return the configuration.